Start defining the "erase" API.
So far, the `erase` extension function has been added to
`ComponentActivity`, `Fragment`, and `View`. One for each of the
supported animation APIs.

This works fine, but my worry is that it will quickly pollute each of
the chosen classes with many `erase` functions.

One benefit though, is that we can choose the correct lifecycles to
observe. For example, we want to observe the `viewLifecycleOwner` within
a `Fragment` and not the `lifecycleOwner`. Inverting the extension
functions may make it difficult to ensure the correct lifecycles are
provided.