Changes to internal and public APIs.
Added support for `AnimatedVectorDrawableCompat` - need to add a message
recommending developers use this on < Android M to prevent memory leaks
(there is no clearListeners function below M for
`AnimatedVectorDrawable`.

Comitted to using the `Activity`, `Fragment` and `View` as receivers for
extension functions. I still believe this causes a lot of duplication
and pollutes the namespace, but it provides the most consistent API. A
little bit of extra work here for a better experience for developers.

Using a lot of implied types, e.g. `Fragment.erase(...) = erase(...)`
might need to rethink that. Purely done to save lines of code.